requirements.txt
Django==1.6.4
South>=0.8.2
Whoosh>=2.5.4
django-crispy-forms>=1.4.0
django-filter>=0.7
django-taggit>=0.12
djangorestframework>=2.3.8
psycopg2>=2.5.1
pycrypto>=2.6
lxml>=3.2.4
requests>=2.0.0
tld>=0.6.1
selenium>=2.37.2
cssselect>=0.9.1
pytest>=2.5.2
pytest-django>=2.6
pytest-cov>=1.6
django-pipeline>=1.3.24
django-settings-context-processor==0.2
django-admin-bootstrapped>=1.6.4
django-sitegate==0.6.1
-e git+https://github.com/audax/python-readability.git@master#egg=python-readability
-e git+https://github.com/audax/django-haystack.git@whoosh-basic-facets#egg=django-haystack
